[ndk-build] Fix paths to toolchain and clang on Windows (#30)
* ndk-build: Windows toolchain binaries are .exe, not wrapper scripts

All toolchain binaries are executable files with a .exe extension. Only
the clang compiler is wrapped in bash scripts (no extension) or batch
scripts (.cmd extension).
Update the extension to match this.

Note that this was tested on NDK r21c, version 21.2.6472646.

* ndk-build: Append extension after ++ in clang++

The ++ was mistakenly appeneded to the full filename of the binary (by
me). On Windows this binary resolves to clang.cmd and ++ is appened to
that for the C++ variant, resulting in clang.cmd++. This binary doesn't
exist, it should be clang++.cmd.

Fixes: 33c00289ecc0a8ba54824c5c71b6a63233163c15