ndk-examples: Remove empty lib.rs in favour of examples
This crate does not contain executables nor libraries, just the examples
in the example directory.  This is already confusing [1] [2] as they can
only be run with something like:

    cargo apk run -p ndk-examples --example hello_world

Or cd-ing into `ndk-examples` before being able to run these.

Unfortunately `cargo apk run -p ndk-examples` keeps working fine,
resulting in:

    Error: Path "target/aarch64-linux-android/debug/libndk_examples.so" doesn't exist.

Despite the crate clearly not exposing a library target.  The same
happens without this change because `cdylib` is not set for `[lib]`,
but time is still spent compiling all the dependencies instead of
erroring out straight away.

[1]: https://github.com/rust-windowing/android-ndk-rs/issues/48#issuecomment-656931548
[2]: https://github.com/rust-windowing/android-ndk-rs/issues/29#issuecomment-635395518