ci: Run all tests in isolation
When using && to cd back out of a subdirectory the error code from cargo
test is implicitly consumed instead of being caught by set -e.  While
this does return a non-zero exit code it does not prevent the rest of
the script from running if more lines follow.

The next lines run and, if successful, do not fail the step because GH
actions base the result of a multiline shell script solely on the last
command [1].

You can try this for yourself by throwing the following three steps in
GH actions and observing the result [2]:

    false
    echo "This is not printed"

    false && echo "This is not printed"

    false && echo "This is not printed"
    echo "Oh noes, his is printed, and the script exits successfully"

In addition structuring the tests this way allows working-directory to
be used, and shows every test individually in a PR to more quickly
identify where things are going wrong.

[1]: https://docs.github.com/en/free-pro-team@latest/actions/reference/workflow-syntax-for-github-actions#exit-codes-and-error-action-preference
[2]: https://github.com/MarijnS95/gh-actions-test/actions/runs/343934038