ndk-macro: Reexport log crate from ndk-glue and use it in macros
In accordance with a prior, identical change for android_logger:
reexport the log crate from ndk-glue and use it in the macro instead of
relying on crate users to provide log in this scope or override the
path.
This is especially useful if a crate wishes to configure the log level
or tag without needing the log crate in the scope where ndk_glue::main
is called.