ndk,ndk-glue: Do not pass Looper ident as userdata pointer
The value of `0` or `1` is not a valid pointer to a userdata structure.
Not that it has to be, but with the nasty consequence that winit Android
code was accidentally reinterpreting this pointer as integer to extract
the identifier from, instead of using the appropriate `ident` field
directly. After that is addressed in winit we can set the pointer back
to 0 cleanly (and prepare for a potential pointer to some user data that
may be passed in the future).

Also add some constants to document these "special" identifiers.