Rework Event type to allow synchronized cleanup
Previously, NativeActivity callbacks returned immediately
after dispatching an asynchronous Event to the user on a UNIX
pipe. This made it impossible to handle destruction events
(`onNativeWindowDestroyed`, `onInputQueueDestroyed`) in
a manner conforming to the Android docs, which state that
all cleanup must be completed *before* the callback returns,
cf. https://developer.android.com/ndk/reference/struct/a-native-activity-callbacks#struct_a_native_activity_callbacks_1a6aaafbbfae4c0ac066b9d61ebb3ab97f

Therefore, some method of blocking and then notifying the callback
thread had to be introduced. In order to accomplish this, I wrote
a simple wrapper on a lockless queue (`crossbeam_queue::SegQueue`)
and an `eventfd` object, which allows sending arbitrary Rust
objects to the user while notifying the Android `Looper` in a
manner preserving the source of the notification.

This in turn enabled a different semantics of the `Event` type.
`Events` now transfer ownership of various Android handles
(e.g. `NativeActivity`, `InputQueue`) to the user. The user is
responsible for storing these handles. When a corresponding
`*Destroyed` event is received, the user must ensure that they
complete all cleanup before dropping the contained `EventSyncGuard`
object and do not touch the relevant handle afterwards.