ci: Don't build target-agnostic ndk docs for all targets
The previous commit tests all documentation for all features and crates
(not just `ndk`). The NDK crate and its documentation build without
issues thanks to the guards introduced by the `"test"` feature. Since
the `ndk` docs do not depend on any target-specific flags (and `ndk-sys`
does not contain any documentation) is is fine to build-test it just
once for the host architecture.