ci: Run target-agnostic build/test steps only once
A lot of these steps - except apk building - are not target-specific (do
not have a `--target` option; `target:` on `actions-rs/toolchain` only
specifies what targets to install). This is a waste of time despite all
duplication running in parallel, and could be used to run a
target-specific apk build in parallel with host-side tests instead.

It should be fine to perform just the apk build step on Windows, and
leave the rest (tests, documentation) to Linux runners only - most of
those steps were not ran on Windows anyway.

This takes CI time down from 6m30s to just under 5m, with around 3m
spent in apk building (in parallel), 3m40s spent testing all cates and
docs on the host, and under 5m for apk building on Windows
(defining the baseline for total build completion time).