Add support for dynamically loaded libraries
Sometimes we need to add .so libraries to the apk that are not linked against the binary, with the intent of loading them dynamically with `dlopen()`.

This commit allows to do that by placing those libraries at the root of the local package, in a `lib/${androideabi}` folder.

The build command now scans for those folders and adds the found .so files to the apk with `aapt`.