ndk/bitmap: Wrap BitmapFormat in mod to hide deprecation warning
This deprecation warning originates from one of the uses by the
`IntoPrimitive`/`TryFromPrimitive` derive macros.  Because these
generate extra code beyond the enum, a `#[allow]` on just the enum won't
cut it.  As we do not want to allow deprecation to slip into the rest of
the module, wrap `BitmapFormat` in a temporary module to hide this
warning and allow clippy in the CI to succeed.

It seems that warnings originating from macro-generated code don't
appear anymore on at least nightly (1.56), but is still present on 1.54.

Fixes https://github.com/rust-windowing/android-ndk-rs/issues/90