ndk/looper: Drop boxed callback after deregistering
The box created in `add_fd_with_callback` is turned into a raw pointer
but its memory is never dropped when the callback requests for
deregistration.

Note that it is still possible to deregister this callback using
`ALooper_removeFd`, where the box cannot be accessed and would thus
leak.  This is likely not a problem in practice because that function is
not exposed in the safe API at the time of writing.
However, it is still possible to replace the callback by calling
`ALooper_addFd` on the same `fd` again, leaking the boxes.