cargo-apk: Use `min_sdk_version` to select compiler target
According to [1] minSdkVersion is used to determine the compiler target
and ultimately limit what API is available at compile-time.  This is
for the most part because using/linking newer API will result in runtime
linker errors on these older platforms.

In other words, using the target sdk version which (as of writing, but
that will change in the near future too) defaults to the highest version
supported by the detected SDK makes it impossible to detect
compatibility issues (symbols that are unavailable) with your app on
older Android APIs despite setting min_sdk_version.

[1]: https://developer.android.com/ndk/guides/sdk-versions#minsdkversion