ndk-macro: Remove faulty crate name fallback in `fn crate_path()`
`proc_macro_crate::crate_name` fails to find our `ndk-glue`
crate when it sits inside `[target.'cfg(target_os =
"android")'.dev-dependencies]` as shown in [RustAudio/cpal#641], with a
fix pending in [bkchr/proc-macro-crate#15].  This function is only ever
called with `"ndk-glue"` which is an invalid identifier and should have
its dash replaced with an underscore to be valid.  However, since this
fallback doesn't seem to have been hit before (we've never received
reports of `` `"ndk-glue"` is not a valid identifier ``) it is safe and
desired to enforce the crate name to reside in `Cargo.toml` and
`proc-macro-crate` being able to find it.

[RustAudio/cpal#641]: https://github.com/RustAudio/cpal/runs/5203107529?check_suite_focus=true
[bkchr/proc-macro-crate#15]: https://github.com/bkchr/proc-macro-crate/pull/15