Read environment variables from `.cargo/config.toml`'s `[env]` section
Environment variables can be set and optionally override the process
environment through `.cargo/config.toml`'s `[env]` section:
https://doc.rust-lang.org/cargo/reference/config.html#env

These config variables have specific precedence rules with regards to
overriding the environment set in the process, and can optionally
represent paths relative to the parent of the containing `.cargo/`
folder.  The entire handling of these variables is implemented in
https://github.com/dvc94ch/cargo-subcommand/pull/12 which can be read
as-is from cargo-apk and ndk-build, through a trait to keep ndk-build
oblivious for cargo-subcommand.