cargo-apk: Export native activity by default on Android S and up
Similar to how we help the user with a default `MAIN` intent filter to
launch the sole `NativeActivity` in their application (and setting a few
other sensible defaults), we should also set Android's new
`android:exported` attribute to `true` if they're targeting Android S
(SDK 31) or higher.

Without this set to `true` the Rust app is not allowed to be started
externally (by ie. a launcher or `adb`) on Android S+, and with only a
single `Activity` currently supported by `ndk-build` it is unlikely (but
not impossible, see shared user IDs etc), that there are other means to
start it internally.

https://developer.android.com/guide/topics/manifest/activity-element#exported