ndk: Remove on_success closure argument from `fn from_status()`
By far the most calls of this function pass a closure that returns the
void type.  All these default calls can be simplified by always
returning the void type in `Result<()>` from `fn from_status()`, and use
the standard `Resuld::map(|()| ...)` to provide a different ok-value in
the few cases an actual return value is desired.