media_codec: Uphold non-`null` guarantees in some erratic scenarios
The Android documentation does not comment on when functions could
possibly return `null`, and while discussion in [#248] finds these
unlikely enough to be communicated back to the user there is no 100%
guarantee and the resulting pointers should as such always be checked
(either at all, or also when `cfg!(debug_assertions)` happens to be
disabled).

Note that this `NonNull::new(..).unwrap()` pattern is found in many
other places in the `ndk`.

[#248]: https://github.com/rust-windowing/android-ndk-rs/pull/248