ndk-glue: Replace `deprecated` annotation on `native_activity()` with doc-comments
It has been quite some time since we introduced this, and our
alternative implementation for more safely passing a native activity to
the user while interoperating cleanly with `ndk_context` still hasn't
found its way.

There are legitimate uses for `native_activity()` that cannot be
performed through the suggested `ndk_context::android_context().vm()`,
making the load of this deprecation warning heavier than the problem it
tries to prevent.

In order to keep some form of documentation for this issue going, it is
now explained in a doc-comment that is also partially replicated on the
other `static` getter functions.