cargo-apk: Append `--target` to blanket `cargo apk --` calls when not provided
When the user doesn't provide a `--target` we default the triple to the
currently connected device over `adb`, or otherwise fall back to
`aarch64`.

While this triple is always used to determine what NDK environment to
provide, it's never added to the arguments passed after `--`; for
example a `cargo apk -- test --no-run` will try to build tests for the
host unless explicitly called with `--target`.