ndk-build,cargo-apk: Rename `intent_filters` back to `intent_filter`
We commonly name vectors of items with their singular form so that they
read nicer in TOML and XML (i.e. `[[intent_filter]]` adds a single entry
to that list, and gets serialized as a single `<intent-filter>` XML
element) and this matches Android's `intent-filter` naming too.

This change in a6f3e13 ("ndk-build: Move default serialization of `MAIN`
intent filter to `cargo-apk` (#241)") was made purely with `Vec` in
mind, but breaks existing manifest parsing nor was anticipated to be a
breaking change in this area as the README still carries `intent_filter`
instead of `intent_filters` in its `Cargo.toml` metadata reference.