ndk: Remove "irrelevant" `hardware_buffer`/`trace` feature in favour of API level (#320)
Thus far feature flags have been used specifically in `ndk-sys` to
control what native libraries are linked to, for content that is
provided outside of the standard linked `libandroid.so`:

https://github.com/rust-windowing/android-ndk-rs/blob/d104d01f116b537babdc94d8c9889581f8d63a02/ndk-sys/src/lib.rs#L43-L53

However, `hardware_buffer` and `trace` are available in `libandroid.so`
directly making these features only guard our Rust bindings.  This is
unnecessarily tedious and inconsistent with the rest of the NDK wrappers
which are only guarded by their minimum API level.