ndk-build: Add an opt-in alternative to adb install
Instead of using adb install, we manually push the APK file to the device.
The location is in /data/local/tmp/.
Then, we manually install the app with pm install~!
After that, we remove the apk from the install location.

SOMETIMES, I ocassionally find that adb install takes too long to install..
This feature is a lovely workaround to that issue. <3

As this is just a workaround for now, it needs to be enabled by adding the CARGO_APK_MANUAL_INSTALL=1 environment variable.

Signed-off-by: Nefo Fortressia <nefothingy@hotmail.com>