ndk-build,cargo-apk: Launch activity directly through `ndk-gdb`
`ndk-gdb` passes `-D` to `am start` to make the process wait until the
debugger is launched and ready, instead of possibly crashing the app
before the debugger can catch it (paired with a nice "waiting for
debugger" dialog and "Force close" button in Android).  While we could
also set the flag ourselves, this change simplifies the workflow on our
end as well now that we don't have to run our own command, potentially
get stuck on `pidof` returning multiple PIDs, and having to override
`no_logcat` (#342).

(For the multiple-PID case though, we can optionally pass `-f` to make
`ndk-gdb` kill the debugger and app first, if this ever becomes an
issue.)