cargo-apk: Inherit package configuration (`version`) from workspace root
Starting with [Rust 1.64] common `[package]` parameters (and
dependencies) can now be specified in the workspace root manifest by
setting `<field>.workspace=true` in a `[package]` and specifying its
value in the workspace root manifest under [`[workspace.package]`].

Since `cargo-apk` reads the `version` field from `[package]` it has to
support this new format and pull the value from the root manifest
instead, in order to support projects utilizing this new format.

This is currently implemented ad-hoc for the version field, but could
be done in a cleaner way if/when more fields are needed.

[Rust 1.64]: https://blog.rust-lang.org/2022/09/22/Rust-1.64.0.html#cargo-improvements-workspace-inheritance-and-multi-target-builds
[`[workspace.package]`]: https://doc.rust-lang.org/cargo/reference/workspaces.html#the-workspacepackage-table