Lock MSRV to 1.60 and validate in CI
Prevent inadvertently bumping MSRV in PRs by checking against MSRV in
the CI, to disallow any large bumps without rationale.  Rust 1.60 is 7
months old at the time of writing, and also required by `winit` which
uses `android-ndk-rs`, hence must be compatible with it.

This does not mean that we cannot bump MSRV at all, but we must remain a
tad conservative; at the same time setting [`rust-version`] in the
manifest provides everyone on Rust 1.56 and higher a helpful error
message when their version is lower than what we require.

[`rust-version`]: https://doc.rust-lang.org/cargo/reference/manifest.html#the-rust-version-field