ndk/asset: Use entire asset length when mapping buffer
According to `AAsset_getBuffer` documentation:

    Get a pointer to a buffer holding the entire contents of the assset.

This means the returned pointer points to the _start of_ the asset, and
will thus be pointing to a memory range as long as the asset.  If the
user had first read X bytes before mapping it as a buffer, there will
now be X bytes cut off at the end of the buffer since
`get_remaining_length()` returns `get_length() - X` at this point.