ndk/media_codec: Wrap common dequeued-buffer status codes in enum
The dequeue API for output buffers can commonly return
`AMEDIACODEC_INFO_OUTPUT_FORMAT_CHANGED` or
`AMEDIACODEC_INFO_OUTPUT_BUFFERS_CHANGED`, both of which are not caught
in `NdkMediaError::from_status()` and show up as an "unknown" error.
Neither really is an error that the user would want to bubble up from
the `Result<>` via `?`, but should rather handle this explicitly.

A similar case already exists for `AMEDIACODEC_INFO_TRY_AGAIN_LATER`
which previously resulted in a `None`, but is now also encapsuled in a
self-documenting manner, for the dequeue_input_buffer()` API as well.