media_error: Add `MediaStatus::from_status_if_negative()` helper
Covers a prevalent `unwrap_err()` pattern for negative status codes,
where we want to use any non-negative status code as a useful value and
otherwise assert that `from_status()` turns the raw value into an
appropriate error code (never into `Ok(())`, which should be unreachable
in the current implementation).

Note that we cannot rely on `from_status()` returning `Ok(())` or
`Err(_)` and translate the `Ok(())` back to the original value, as it
handles (as of writing) two positive error codes which only relate to a
specific `AMediaCodec` API, and any unrecognized codes including >0 are
treated as `UnknownStatus`.

The function consumes `Into<isize>` instead of `i32/c_int` or
`ffi::media_status_t` to support passing through numbers from various
APIs that return 64-bit numbers (e.g. `MediaCodec`).  The value only
needs to be truncated (with a panic on `try_into()`) when it is lower
than zero and needs to be converted to `ffi::media_status_t`.