bitmap: Strip off repeated `Android` prefix from structs and enums
This convention is already applied to `(Android)BitmapError` and
`(Android)BitmapFormat`, and to many other `A`-prefixed structs
elsewhere in the `ndk` crate: there is no need to repeat the Android
prefix or letter when describing strucures and enumerations.

Also strip off the `Bitmap` prefix in a `Result` alias, which isn't
common either when only filling in a default for the generic `Error`
argument.