ndk/media_format: Implement `Debug` in terms of `Display`
Currently `Debug` only prints a raw pointer which is rather useless to
look at.  In addition to the pointer, also print the `Display`
representation of `MediaFormat` which uses Android's `toString()`
function to create a human-readable string of the various fields set
inside of it.

On a side-note it is "great" to see that `toString()` is not currently
available as a user function via a lifetimed `CStr` as it is invalidated
in a nontrivial way (e.g. when `toString()` is called again, and
possibly also when `set_()` functions are called which are not currently
borrowing it mutably).