ndk/media_format: Make all mutating functions take `self` by `&mut`
All setters can inherently change the internals of the format, and should
furthermore invalidate - or not be allowed while - borrowed strings or
buffers (are) acquired from it.

Unforatunately [`getString()` is documented as]:

    The returned string is owned by the format, and remains valid until
    the next call to getString, or until the format is deleted.

Meaning that it is not reentrant, and invalidates any data previously
acquired from `str()`; hence this getter is marked as `&mut self` too.
The same is however not the case for `getBuffer()`.

[`getString()` is documented as]: https://github.com/rust-mobile/ndk/issues/451