ndk: Requre all off-thread `dyn Fn*` callback types to implement `Send`
In all these cases the NDK either documents or implements the callback
to be invoked from a single, separate thread.  For this to be allowed
by Rust thread safety, the closures and their (moved) contents are
effectively  "moved" to a different thread (`Send`) but not accessed
concurrently (`Sync`), but the type definitions were never requiring
this marker trait which could lead to undefined/invalid behaviour.

In addition some `Box`ed callbacks may have been dropped before a new
callback is passed to the NDK, leading to a possible race condition.
By storing the new callback _after_ registering it with the NDK, the
previous callback is now dropped later to ensure such race conditions
no longer happen (assuming AOSP callback invocations are properly
synchronized with the setters).

Multiple requests have been filed to solidify this "behavioural
contract" in the documentation where not done so yet:
https://issuetracker.google.com/issues/300602767#comment12
https://issuetracker.google.com/issues/318944941