image_reader: Add `AImageReader_newWithDataSpace()` from `api-level-34`
Android 14 (U, API level 34) adds `DataSpace` support to `ImageReader`
with a new constructor and a getter, and finally switches to the regular
`HardwareBufferFormat` instead of the restricted `ImageFormat` enum
to select the underlying buffer format.  After all, an `AImage` can be
turned into a `HardwareBuffer` which supports a wider range of formats.

Also clean up and commonize some panics on pedantic `.try_from()` checks.