native_window: Use `i8` in `FrameRate` `enum`s to match functions
All functions that deal with these types take an `i8`; just the
underlying `ndk-sys` brindings are misrepresented because `bindgen` does
not know what integer type to use for an `enum`.

Using an `i8` makes casting more natural, without unnecessary overflow
checks.  The compiler will make sure that the constants assigned to this
enum never exceed `i8`, and the NDK would never overflow these constants
in the first place as that breaks the existing bindings.

Note that these types will be used with the same `i8` type in the
`SurfaceControl`/`SurfaceTransaction` bindings.