ndk/image_reader: Special-case return statuses in `Image`-acquire functions
Both async and non-async `acquire_next/latest_image()` functions will
return `ImgreaderNoBufferAvailable` when the producer has not provided
a buffer that is either ready for consumption or that can be blocked
on (either inside a non-async method, or by returning the accompanying
fence file descriptor).  But only the non-`_async()` functions were
marked as if this is a common case by returning an `Option<>`, seemingly
out of the assumption that the `_async()` functions can _always_ give
you an image (if the `MaxImagesAcquired` limit is not reached) but with
a file-descriptor sync fence to wait on.  This is not the case as the
producer needs to submit a buffer together with a sync fence on the
producer-end first.

Hence the current API signatures create the false assumption that only
non-async functions can "not have a buffer available at all", when
the exact same is true for `_async()` functions, in order to provide
an image buffer with a fence that is signalled when it is ready for
reading.

Instead of special-casing this error in the `_async()` functions,
special-case both `NoBufferAvailable` and `MaxImagesAcquired` in a new
`enum AcquireResult` and let it be returned by both non-async and async
functions.