ndk/event: Implement `SourceClass` `bitflag` and provide `Source::class()` getter
This `enum` was never `pub` even though it serves a slight purpose to
more easily classify various event sources, which is accounted for in
the bitmask-like value of `AINPUT_SOURCE_XXX`.  Even though every source
currently only appears to use a single class value, the class variants
all describe an individual bit and should be treated as a bitflags type.